Ensure user provided JSX gets highlighted.
This should not be considered a stable fix.
The `monaco-jsx-highlighter` package is in
a state of flux (essentially, it's still in beta)
and may break in future.

TODO: That package has now hit a 1.x release
and at some point in the future we should see if
that is stable enough to move to.